chore: provide single recommended demo
