+++ .babelrc
@@ -4,7 +4,7 @@
       "env",
       {
         "targets": {
-          "node": "4"
+          "node": "8"
         }
       }
     ]
