--- example-complete/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    [
-      "env",
-      {
-        "modules": false,
-        "targets": { "browsers": "last 2 versions" }
-      }
-    ],
-    "react"
-  ]
-}
