--- example-experimental/Dockerfile
@@ -1,17 +1,0 @@
-FROM node:8.9-alpine as base
-LABEL maintainer="hedgecock.d@gmail.com"
-
-WORKDIR /usr/src/app
-
-# Install app dependencies
-# A wildcard is used to ensure both package.json AND package-lock.json are copied
-COPY example-experimental/package*.json ./
-RUN npm install
-
-RUN npm install -g serve
-
-# Bundle app source
-COPY example-experimental .
-
-EXPOSE 5000
-EXPOSE 3000
