--- example-experimental/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "experimental-feature-set-example-configuration",
-  "version": "0.0.0",
-  "description": "Demo configurations for using the experimental feature set",
-  "scripts": {
-    "start": "webpack-dev-server --env=development --progress",
-    "build": "NODE_ENV=production webpack --env=production --progress"
-  },
-  "license": "ISC",
-  "dependencies": {
-    "ionicons": "^3.0.0",
-    "prop-types": "^15.6.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
-  },
-  "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "html-webpack-plugin": "^3.0.7",
-    "svg-symbol-sprite-loader": "^0.3.2",
-    "webpack": "^4.1.1",
-    "webpack-cli": "2.0.12",
-    "webpack-dev-server": "^3.1.1"
-  }
-}
