--- example-experimental/public/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-  <!-- The very amazing polyfill.io -->
-  <script src="https://ft-polyfill-service.herokuapp.com/v2/polyfill.min.js"></script>
-</head>
-
-<body>
-  <div id="root"></div>
-</body>
-
-</html>
